Fix: Some mesh points would be outside of the velocity model
Some mesh points are not just slightly outside of the velocity model, but have a significant distance from the boundary of the
region. To have a more consistent solution, all the mesh points outside the velocity model will fall back to using CVM or 1D layer
profile.
This is especially important for higher frequency simulations since the element size will be very small and the whole element might
be outside of the region and cause troubles and errors.