fix: Three small errors have been fixed
The `minPointZ` value in `drm_planewaves.c` should be a double-precision floating number instead of an integer.
Correct the edge cases when determining whether the mesh point is within the 3D velocity model.
Qs and Qp won't be recalculated anymore if the mesh point is not within the 3D velocity model (and thus uses 1D layer profile).