feat: The path to the 3D Velocity Model is now configurable
The path to the 3D Velocity Model is now configurable.
Properly broadcast path strings.
The default value is set for the checkpoint path.
The plane input file is now an optional parameter, and the default value is the same as the parameter input file (read it directly from the command line).
Fix the issue that the ``source_directory_output`` is not correctly read (move the section above fclose(fp);)