fix: The `plotmesh.m` MATLAB script works again
Due to the previous changes to the mesh output binary files, the old `plotmesh.m` did not work.
Now it works again, and the input it needs is also simplified. Domain dimensions are not needed anymore.
The example input `plotmeshinput.in` for using `plotmesh.m` has also been updated accordingly.